Continuing renaming of solution.
