+++ CircuitBreaker.Tests/Properties/AssemblyInfo.cs
