+++ CircuitBreaker/ICircuitBreaker.cs
