+++ CircuitBreaker/Properties/Resource.resx
