+++ CircuitBreaker/SimpleCircuitBreaker.nuspec
